refactor code and tests

+++ lib/inspec/fetcher/git.rb
@@ -68,20 +68,19 @@ module Inspec::Fetcher
       else
         Dir.mktmpdir do |working_dir|
           checkout(working_dir)
-          if Dir.empty?(working_dir)
-            # If the temporary working directory is empty after checkout,
-            # this means the git repository did not contain any files (or the checkout failed).
-            # In this case, remove the destination directory to avoid
-            # leaving an empty or invalid profile directory.
-            Inspec::Log.debug("Remove #{destination_path} directory because temp directory {working_dir} it is empty")
-            FileUtils.rm_r(destination_path)
-          elsif @relative_path
+          if @relative_path
             perform_relative_path_fetch(destination_path, working_dir)
           else
             Inspec::Log.debug("Checkout of #{resolved_ref.nil? ? @remote_url : resolved_ref} successful. " \
                                 "Moving checkout to #{destination_path}")
             FileUtils.cp_r(working_dir + "/.", destination_path)
           end
+          # If the temporary working directory is empty after checkout,
+          # this means the git repository did not contain any files (or the checkout failed).
+          # In this case, remove the destination directory to avoid
+          # leaving an empty or invalid profile directory.
+          Inspec::Log.debug("Remove #{destination_path} directory because temp directory #{working_dir} it is empty")
+          FileUtils.rm_r(destination_path) unless Dir.empty?(working_dir) || Dir.empty?(working_dir)
         end
       end
       @repo_directory
